feat: implement DatePicker component and integrate it into Home screen
