+++ app/(protected)/home/index.tsx
@@ -5,9 +5,12 @@ import { styles } from "./style";
 import { getAuth } from "firebase/auth";
 import { getCurrentDate } from "@/utils/getCurrentData";
 import TransactionsList from "@/components/transactions/list";
+import { DatePicker } from "../../../components/datePicker";
+import { useState } from "react";
 
 export default function Home() {
   const auth = getAuth();
+  const [date, setDate] = useState(new Date());
 
   return (
     <View style={styles.container}>
@@ -21,6 +24,11 @@ export default function Home() {
         <SummaryCard value={5000} type="income" />
         <SummaryCard value={2000} type="outcome" />
       </View>
+      <DatePicker value={date} onChange={(data) => {
+        setDate(data);
+      }}
+                  label="Data">
+      </DatePicker>
       <View style={styles.transactions}>
         <TransactionsList />
       </View>
